DashboardFragment logic & LoginActivity logic
