+++ app/src/main/AndroidManifest.xml
@@ -7,11 +7,13 @@
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@drawable/logo3"
-        android:label="@string/app_name"
+        android:label="UFS CLINIC"
         android:supportsRtl="true"
         android:theme="@style/Theme.Material3.Dark.NoActionBar"
         tools:targetApi="31">
-
+        <activity
+            android:name=".Dashboard"
+            android:exported="false" />
         <activity
             android:name=".Splash"
             android:exported="true">
