+++ app/src/main/res/layout/activity_dashboard.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:padding="20dp"
     android:background="@color/white"
-    tools:context=".DashboardFragment">
+    tools:context=".Dashboard">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -39,6 +39,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:src="@drawable/avatar"
+                android:contentDescription="Profile picture"
                 android:outlineAmbientShadowColor="@color/blue"/>
         </androidx.cardview.widget.CardView>
     </LinearLayout>
@@ -67,6 +68,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="match_parent"
                     android:src="@drawable/ic_search"
+                    android:contentDescription="Search icon"
                     android:layout_gravity="center"
                     android:padding="8dp"/>
 
@@ -86,6 +88,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="match_parent"
                     android:src="@drawable/ic_mic"
+                    android:contentDescription="Mic icon"
                     android:layout_gravity="center"
                     android:padding="8dp"/>
             </LinearLayout>
